src/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
paho-mqtt==1.6.1
pydub==0.25.1
pytube==15.0.0
redis==5.0.1
requests==2.31.0
six==1.16.0
spotipy==2.23.0
urllib3==2.1.0